requirements.txt
Pillow==9.3.0
wavedrom==2.0.3.post3
CairoSVG==2.5.2
json5==0.9.10